Split a pdf into documents and saving documents as different pdfswith ability to set split size and input/output file names
